+++ pom.xml
@@ -34,6 +34,7 @@
         <snakeyaml.version>1.33</snakeyaml.version>
         <structs.version>324.va_f5d6774f3a_d</structs.version>
         <caffeine-api.version>2.9.3-65.v6a_47d0f4d1fe</caffeine-api.version>
+        <instance-identity.version>116.vf8f487400980</instance-identity.version>
     </properties>
 
     <name>UiPath Plugin</name>
@@ -215,7 +216,7 @@
         <dependency>
             <groupId>org.jenkins-ci.modules</groupId>
             <artifactId>instance-identity</artifactId>
-            <version>116.vf8f487400980</version>
+            <version>${instance-identity.version}</version>
         </dependency>
         <dependency>
             <groupId>org.yaml</groupId>
