+++ src/main/java/com/uipath/uipathpackage/UiPathAssets.java
@@ -102,10 +102,10 @@ public class UiPathAssets extends Builder implements SimpleBuildStep {
         util.validateRuntime(launcher);
 
         try {
-            EnvVars envVars = TaskScopedEnvVarsManager.selectOnlyRequiredEnvironmentVariables(run, env, listener);
+            EnvVars envVars = TaskScopedEnvVarsManager.addRequiredEnvironmentVariables(run, env, listener);
 
             CliDetails cliDetails = util.getCliDetails(run, listener, envVars, launcher);
-            String buildTag = envVars.get("BUILD_TAG");
+            String buildTag = envVars.get(EnvironmentVariablesConsts.BUILD_TAG);
 
             FilePath expandedCsvFilePath = filePath.contains("${WORKSPACE}") ?
                     new FilePath(launcher.getChannel(), envVars.expand(filePath)) :
