+++ src/main/java/com/uipath/uipathpackage/UiPathDeploy.java
@@ -205,10 +205,10 @@ public class UiPathDeploy extends Recorder implements SimpleBuildStep {
         util.validateRuntime(launcher);
 
         try {
-            EnvVars envVars = TaskScopedEnvVarsManager.selectOnlyRequiredEnvironmentVariables(run, env, listener);
+            EnvVars envVars = TaskScopedEnvVarsManager.addRequiredEnvironmentVariables(run, env, listener);
 
             CliDetails cliDetails = util.getCliDetails(run, listener, envVars, launcher);
-            String buildTag = envVars.get("BUILD_TAG");
+            String buildTag = envVars.get(EnvironmentVariablesConsts.BUILD_TAG);
 
             FilePath expandedPackagePath = packagePath.contains("${WORKSPACE}") ?
                     new FilePath(launcher.getChannel(), envVars.expand(packagePath)) :
