+++ src/main/java/com/uipath/uipathpackage/UiPathRunJob.java
@@ -287,10 +287,10 @@ public class UiPathRunJob extends Recorder implements SimpleBuildStep {
         util.validateRuntime(launcher);
 
         try {
-            EnvVars envVars = TaskScopedEnvVarsManager.selectOnlyRequiredEnvironmentVariables(run, env, listener);
+            EnvVars envVars = TaskScopedEnvVarsManager.addRequiredEnvironmentVariables(run, env, listener);
 
             CliDetails cliDetails = util.getCliDetails(run, listener, envVars, launcher);
-            String buildTag = envVars.get("BUILD_TAG");
+            String buildTag = envVars.get(EnvironmentVariablesConsts.BUILD_TAG);
 
             JobOptions jobOptions = new JobOptions();
             if (cliDetails.getActualVersion().supportsNewTelemetry()) {
