+++ src/main/java/com/uipath/uipathpackage/models/LocalizationOptions.java
@@ -1,9 +1,8 @@
 package com.uipath.uipathpackage.models;
 
-public abstract class LocalizationOptions extends TelemetryOptions {
-    private String language;
+public interface LocalizationOptions {
 
-    public String getLanguage() { return language; }
+    public String getLanguage();
 
-    public void setLanguage(String language) { this.language = language; }
+    public void setLanguage(String language);
 }
