+++ src/main/java/com/uipath/uipathpackage/util/EnvironmentVariablesConsts.java
@@ -1,0 +1,11 @@
+package com.uipath.uipathpackage.util;
+
+public final class EnvironmentVariablesConsts {
+    private EnvironmentVariablesConsts() {
+        // Prevent instantiation
+    }
+
+    public static final String BUILD_TAG = "BUILD_TAG";
+    public static final String WORKSPACE = "WORKSPACE";
+    public static final String UIPATH_EXTENSIONS_CLI_TELEMETRY_ENABLED = "UIPATH_EXTENSIONS_CLI_TELEMETRY_ENABLED";
+}
