+++ src/main/java/com/uipath/uipathpackage/util/TaskScopedEnvVarsManager.java
@@ -8,11 +8,14 @@ import javax.annotation.Nonnull;
 import java.io.IOException;
 
 public class TaskScopedEnvVarsManager {
-    public static EnvVars selectOnlyRequiredEnvironmentVariables(@Nonnull Run<?, ?> run, @Nonnull EnvVars env, @Nonnull TaskListener listener) throws IOException, InterruptedException {
+    public static EnvVars addRequiredEnvironmentVariables(
+            @Nonnull Run<?, ?> run,
+            @Nonnull EnvVars env,
+            @Nonnull TaskListener listener) throws IOException, InterruptedException {
         EnvVars envVars = run.getEnvironment(listener);
-        envVars.put("WORKSPACE", env.get("WORKSPACE"));
-        if (env.get("UIPATH_EXTENSIONS_CLI_TELEMETRY_ENABLED") != null)
-            envVars.put("UIPATH_EXTENSIONS_CLI_TELEMETRY_ENABLED", env.get("UIPATH_EXTENSIONS_CLI_TELEMETRY_ENABLED"));
+        envVars.put(EnvironmentVariablesConsts.WORKSPACE, env.get(EnvironmentVariablesConsts.WORKSPACE));
+        if (env.get(EnvironmentVariablesConsts.UIPATH_EXTENSIONS_CLI_TELEMETRY_ENABLED) != null)
+            envVars.put(EnvironmentVariablesConsts.UIPATH_EXTENSIONS_CLI_TELEMETRY_ENABLED, env.get(EnvironmentVariablesConsts.UIPATH_EXTENSIONS_CLI_TELEMETRY_ENABLED));
 
         return envVars;
     }
